+count products int64 to int

+++ models/product.go
@@ -111,8 +111,8 @@ func (ps *ProductService) FindProductsByBarcode(barcodeStr string) ([]Product, e
 }
 
 // GetProducts возвращает список продуктов
-func (ps *ProductService) GetProducts(offset int, limit int) ([]Product, int64, error) {
-	var count int64
+func (ps *ProductService) GetProducts(offset int, limit int) ([]Product, int, error) {
+	var count int
 
 	sqlProd := "SELECT p.id, p.name, p.manufacturer_id, m.name FROM products p " +
 		"		LEFT JOIN manufacturers m ON p.manufacturer_id = m.id" +
